Update Filters on Resource Table
