+++ changelogs/unreleased/742-resource-table-filter.yml
@@ -1,0 +1,6 @@
+description: Update Filters on Resource Table
+issue-nr: 742
+change-type: minor
+destination-branches: [master, iso4]
+sections:
+  feature: "{{description}}"
