+++ src/UI/Pages/Resources/TableControls/index.ts
@@ -1,0 +1,1 @@
+export * from "./ResourceTableControls";
