+++ src/UI/words.ts
@@ -199,10 +199,10 @@ const dict = {
   "resources.column.value": "Value",
   "resources.column.numberOfDependencies": "Number of dependencies",
   "resources.column.deployState": "Deploy State",
-  "resources.filters.status.placeholder": "Select a Deploy State...",
-  "resources.filters.agent.placeholder": "Search for an agent...",
-  "resources.filters.value.placeholder": "Search for a value...",
-  "resources.filters.type.placeholder": "Search for a type...",
+  "resources.filters.status.placeholder": "Deploy State...",
+  "resources.filters.agent.placeholder": "Agent...",
+  "resources.filters.value.placeholder": "Value...",
+  "resources.filters.type.placeholder": "Type...",
   "resources.details.title": "Details",
   "resources.details.id": "Id",
   "resources.details.lastDeploy": "Last Deploy",
